% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{add_analysis}
\alias{add_analysis}
\title{Add analysis to tapestri object. the analysis should have same # of cells as tapestri object}
\usage{
add_analysis(tapestri_object, assay_name, analysis_name, new_data)
}
\arguments{
\item{tapestri_object}{tapestri object to update}

\item{assay_name}{feature belongs to assay variants, cnv, proteins}

\item{analysis_name}{}

\item{new_data}{}
}
\value{

}
\description{
Add analysis to tapestri object. the analysis should have same # of cells as tapestri object
}
