% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiomics_object.R
\name{add_assay}
\alias{add_assay}
\title{Add Assay Object to Multiomics object (moo)}
\usage{
add_assay(moo, assay, keep_common_cells = FALSE)
}
\arguments{
\item{moo}{Multiomics object}

\item{assay}{new Assay object to add}

\item{keep_common_cells}{default FALSE. Will throw error if number of cells and cell ids dont match. If TRUE merge assays and only keep intersect of cells from all assays.}
}
\value{

}
\description{
Add Assay Object to Multiomics object (moo)
}
