% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{add_feature}
\alias{add_feature}
\title{Add features to tapestri object}
\usage{
add_feature(tapestri_object, assay_name, feature_name, new_data)
}
\arguments{
\item{tapestri_object}{tapestri object to update}

\item{assay_name}{feature belongs to assay variants, cnv, proteins}

\item{feature_name}{name of feature}

\item{new_data}{new date to add to taesptri object}
}
\value{
updated tapestri object
}
\description{
Add features to tapestri object
}
