% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate Variants based on variant ID field}
\usage{
annotate_variants(variant_ids)
}
\arguments{
\item{variant_ids}{variant id like 'ASXL1:chr20:31022959:T/C'}
}
\value{

}
\description{
Annotate Variants based on variant ID field
}
