% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay_object.R
\docType{class}
\name{Assay-class}
\alias{Assay-class}
\alias{Assay}
\title{Assay class}
\description{
Store assay specific details for given sample
}
\section{Slots}{

\describe{
\item{\code{assay_name}}{}

\item{\code{cell_annotations}}{}

\item{\code{feature_annotations}}{}

\item{\code{data_layers}}{}

\item{\code{analysis_layers}}{}
}}

