% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_normalization.R
\name{clr_by_cell}
\alias{clr_by_cell}
\title{CLR normalization by cell}
\usage{
clr_by_cell(counts)
}
\arguments{
\item{counts}{counts data}
}
\value{
normalized counts
}
\description{
CLR normalization by cell
}
\examples{
\dontrun{
counts.norm <- clr_by_feature(counts)
}
}
