% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay_object.R
\name{create_assay}
\alias{create_assay}
\title{Create Tapestri_Assay Object}
\usage{
create_assay(assay_name, cell_annotations, feature_annotations)
}
\arguments{
\item{assay_name}{assay name (dna, cnv, protein)}

\item{cell_annotations}{table of data annotating cells, including sample labels}

\item{feature_annotations}{table of data annotating features. i.e. variant names or amplicon name}
}
\value{
Tapestri_Assay object
}
\description{
Create Tapestri_Assay Object
}
