% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiomics_object.R
\name{create_moo}
\alias{create_moo}
\title{Create Multiomics Object}
\usage{
create_moo(experiment_name, cell_annotations)
}
\arguments{
\item{data}{data}

\item{type}{assay type}

\item{name}{sample name}
}
\value{
Analyte Analyte object
}
\description{
Create Multiomics Object
}
\examples{
\dontrun{
genotypes <- extract_genotypes(loom, barcodes, gt.mask=TRUE)
assay <- create_assay(genotypes, "snv", "sample")
}
}
