% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_subclones.R
\name{define_subclones}
\alias{define_subclones}
\title{define subclones based on NGT values}
\usage{
define_subclones(assay, small_subclone = 0.01, ignore_zygosity = FALSE)
}
\arguments{
\item{assay}{Assay object for DNA variants}

\item{small_subclone}{threshold (\%) for considering subclones too small}

\item{ignore_zygosity}{collapse HET/HOM to just mutated}
}
\value{

}
\description{
Similar to how Tapestri Insights v2.2 calculates subclones
}
