% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_variants}
\alias{filter_variants}
\title{Filter raw genotypes based on quality}
\usage{
filter_variants(
  tapestri_object,
  gqc = 30,
  dpc = 10,
  afc = 20,
  mv = 50,
  mc = 50,
  mm = 1,
  gt.mask = FALSE
)
}
\arguments{
\item{gqc}{Genotype quality cutoff (default 30)}

\item{dpc}{Read depth cutoff (default 10)}

\item{afc}{Allele frequency cutoff (default 20)}

\item{mv}{Remove variants with < mv of known values (default 50)}

\item{mc}{Remove variants with < mc of known values (default 50)}

\item{mm}{Remove variants mutated in < mm of cells (default 1)}

\item{gt.mask}{mask low quality GT as missing (if GQ/DP/AF lower than cutoff, default FALSE)}

\item{loom}{Loom file}
}
\value{
Filtered genotypes
}
\description{
Filter raw genotypes based on quality
}
