% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{h5_reader}
\alias{h5_reader}
\title{Read Tapestri multi-omics h5 file}
\usage{
h5_reader(filename, min_mutation_rate = 0.01)
}
\arguments{
\item{filename}{h5 file}

\item{min_mutation_rate}{only read variants with mutations rate higher then treshold. Primarly done to reduce size of data in memory}
}
\value{
Tapestri multi-omics object
}
\description{
Read Tapestri multi-omics h5 file
}
\examples{
\dontrun{
tapestri_raw = h5_reader(filename,min_mutation_rate = 0.1)
}
}
