% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_normalization.R
\name{log_colMeans}
\alias{log_colMeans}
\title{Log normalization factor per feature}
\usage{
log_colMeans(x)
}
\arguments{
\item{x}{Sample name}
}
\value{
normalization factor
}
\description{
Log normalization factor per feature
}
\examples{
\dontrun{
norm_factor <- log_colMeans(data)
}
}
