% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_normalization.R
\name{log_norm_by_cell}
\alias{log_norm_by_cell}
\title{Log normalization by cell}
\usage{
log_norm_by_cell(counts)
}
\arguments{
\item{counts}{counts data}

\item{...}{}
}
\value{
normalized counts
}
\description{
Log normalization by cell
}
\examples{
\dontrun{
counts.norm <- log_norm_by_cell(counts)
}
}
