% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_normalization.R
\name{log_norm_by_feature}
\alias{log_norm_by_feature}
\title{Log normalization by feature}
\usage{
log_norm_by_feature(counts)
}
\arguments{
\item{counts}{counts data}
}
\value{
normalized counts
}
\description{
Log normalization by feature
}
\examples{
\dontrun{
counts.norm <- log_norm_by_feature(counts)
}
}
