% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{multiassay_heatmap}
\alias{multiassay_heatmap}
\title{Plot heatmap}
\usage{
multiassay_heatmap(moo, cluster)
}
\arguments{
\item{cluster}{cluster labels}

\item{analytes}{multi assay object}

\item{snv_features}{snv features}

\item{prot_features}{prot features}

\item{groupby}{group by gene/chr and select variant with highest mutation rate}
}
\value{
heatmap object
}
\description{
Plot heatmap
}
\examples{
\dontrun{
p <- multiassay_heatmap(analytes, cluster, snv_features=c("DNMT3A", "JAK2"),
   prot_features=c("CD3", "CD34"), groupby="gene")
}
}
