% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TapestriR_object.R
\docType{class}
\name{Multiomics-class}
\alias{Multiomics-class}
\alias{Multiomics}
\title{The Multiomics Class}
\description{
Store multiomics-multisample data generated from various Tapestri pipelines
}
\section{Slots}{

\describe{
\item{\code{experiment_name}}{Experiment Name}

\item{\code{dna}}{DNA assay}

\item{\code{cnv}}{read counts}

\item{\code{protein}}{Protein assay}

\item{\code{cell_annotations}}{cell annotations including sample labels}
}}

