% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiomics_object.R
\name{names.Tapestri_Multiomics}
\alias{names.Tapestri_Multiomics}
\title{get names of object}
\usage{
\method{names}{Tapestri_Multiomics}(x)
}
\arguments{
\item{x}{Tapestri_Multiomics object}
}
\description{
get names of object
}
