% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize raw counts counts}
\usage{
normalize_data(analyte, group_by, method)
}
\arguments{
\item{analyte}{Analyte object to be normalized (Supported analytes : prot/dna)}

\item{group_by}{normalization to be performed across rows or columns (Supported : cell/feature)}

\item{method}{Normalization method (Supported : mean/median/clr/sum)}
}
\value{
Analyte Analyte object with normalized data
}
\description{
Normalize raw counts counts
}
\examples{
\dontrun{
dna_analyte <- normalize_data(dna_analyte, group_by="cell", method = "median")
prot_analyte <- normalize_data(prot_analyte, group_by="cell", method = "clr")
}
}
