% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_reads_normalization.R
\name{normalize_dna_reads}
\alias{normalize_dna_reads}
\title{Normalize barcode data}
\usage{
normalize_dna_reads(reads, advanced = FALSE)
}
\arguments{
\item{reads}{matrix of read counts per amplicon per cell}

\item{advanced}{advance normalization (dividing read counts by cell median before applying normalization by total sum)}
}
\value{
normalized_barcodes - normalized barcode data
}
\description{
Normalize barcode data
}
