% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_assay_h5}
\alias{read_assay_h5}
\title{Read Tapestri multi-omics h5 file}
\usage{
read_assay_h5(filename, assay_name, min_mutation_rate = 0.01)
}
\arguments{
\item{filename}{h5 file path}

\item{assay_name}{name of assay to load from h5}

\item{min_mutation_rate}{only read variants with mutations rate higher then treshold. Primarly done to reduce size of data in memory}
}
\value{
Tapestri multi-omics object
}
\description{
Read Tapestri multi-omics h5 file
}
\examples{
\dontrun{
tapestri_raw = h5_reader(filename,min_mutation_rate = 0.1)
}
}
