% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_insights_export}
\alias{read_insights_export}
\title{Read variant assay from data exported from Tapestri Insights}
\usage{
read_insights_export(export_dir)
}
\arguments{
\item{export_dir}{directory path to exported data. Should contain "AF.csv", "DP.csv", "GQ.csv", "NGT.csv", "README.txt", "Variants.csv"}
}
\value{

}
\description{
Creates a Assay object from Tapestri Insights exported zip file. Inlcudes all the DNA variant variant layers, plus variant annotations. If subclone information exists creates any analysis layer with the subclone labels.
}
