% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_loom}
\alias{read_loom}
\title{Read loom file created by Tapestri pipeline}
\usage{
read_loom(filename, min_mutation_rate = 0.05)
}
\arguments{
\item{filename}{loom file path}

\item{min_mutation_rate}{only read variants with mutations rate higher then treshold. Primarly done to reduce size of data in memory}
}
\value{
Tapestri object
}
\description{
Read loom file created by Tapestri pipeline
}
\examples{
\dontrun{
tapestri_raw = loom_reader(filename,min_mutation_rate = 0.1)
}
}
