% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_tap}
\alias{read_tap}
\title{read a tap file generated by Tapestri pipeline}
\usage{
read_tap(filename, experiment_name = NA)
}
\arguments{
\item{filename}{path to tap file}

\item{experiment_name}{name of experiment, if none file name is used}
}
\value{
multiomics object
}
\description{
read a tap file generated by Tapestri pipeline
}
