% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{recode_genotypes}
\alias{recode_genotypes}
\title{Recode NGT values HET, HOM, WT or MUT/WT if zygosity is ignored}
\usage{
recode_genotypes(x, collapse_zygosity = TRUE)
}
\arguments{
\item{x}{NGT vector}

\item{collapse_zygosity}{default TRUE (WT,MUT). if FALSE (WT, HET, HOM)}
}
\value{

}
\description{
Recode NGT values HET, HOM, WT or MUT/WT if zygosity is ignored
}
