% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay_object.R
\name{subset_assay}
\alias{subset_assay}
\title{Title}
\usage{
subset_assay(assay, keep_cell_ids = TRUE, keep_feature_ids = TRUE)
}
\arguments{
\item{assay}{}

\item{barcodes}{}
}
\value{

}
\description{
Title
}
