% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiomics_object.R
\docType{class}
\name{Tapestri_Multiomics-class}
\alias{Tapestri_Multiomics-class}
\alias{Tapestri_Multiomics}
\title{The Tapestri_Multiomics Class}
\description{
Store Tapestri_Multiomics-multisample data generated from various Tapestri pipelines
}
\section{Slots}{

\describe{
\item{\code{cell_annotations}}{cell annotations including sample labels}

\item{\code{assays}}{}

\item{\code{metadata}}{}
}}

