% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tapestri_ploidy_plot}
\alias{tapestri_ploidy_plot}
\title{Title}
\usage{
tapestri_ploidy_plot(normalized_reads, clusters)
}
\arguments{
\item{normalized_reads}{read counts per amplicon per cell}

\item{clusters}{cluster labels}
}
\value{

}
\description{
Title
}
