% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tapestri_scatterplot}
\alias{tapestri_scatterplot}
\title{Scatter plot wrapper for ggplots that makes plotting Tapestri object easier}
\usage{
tapestri_scatterplot(x, y, color_by)
}
\arguments{
\item{x}{plot of x}

\item{y}{plot of y}

\item{color_by}{what features of color by}
}
\value{
ggplot object
}
\description{
Scatter plot wrapper for ggplots that makes plotting Tapestri object easier
}
