% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tapestri_violinplot}
\alias{tapestri_violinplot}
\title{violin plot wrapper for ggplots that makes plotting Tapestri object easier}
\usage{
tapestri_violinplot(clusters, features)
}
\arguments{
\item{clusters}{clusters to split data by}

\item{features}{features to plot as violin graph}
}
\value{

}
\description{
violin plot wrapper for ggplots that makes plotting Tapestri object easier
}
