% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{umap_theme}
\alias{umap_theme}
\title{Simple theme to show axis-less projection of umaps}
\usage{
umap_theme()
}
\value{
ggplot theme
}
\description{
Simple theme to show axis-less projection of umaps
}
