% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{validate_variant_annotations}
\alias{validate_variant_annotations}
\title{Annotate Variants based on variant ID field}
\usage{
validate_variant_annotations(variant_annotations)
}
\arguments{
\item{variant_annotations}{table of annotations. unique id column must exist. if its in format 'SF3B1:chr2:198266943:C/T' try to parse into CHROM and POS}
}
\value{

}
\description{
Annotate Variants based on variant ID field
}
