% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay_object.R
\name{add_data_layer}
\alias{add_data_layer}
\title{Add additional layers of data to Assay Object}
\usage{
add_data_layer(assay, layer_name, data)
}
\arguments{
\item{assay}{Assay object to add data to}

\item{layer_name}{name of layer}

\item{data}{new data to add}
}
\value{
Assay Object
}
\description{
Add additional layers of data to Assay Object
}
