% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_reads_normalization.R
\name{compute_ploidy}
\alias{compute_ploidy}
\title{Identify ploidy for samples}
\usage{
compute_ploidy(reads, clusters, baseline_cluster)
}
\arguments{
\item{bcfile.norm}{Normalised barcode distribution}

\item{baseline_cluster_cells}{vector of TRUE/FALSE of size number of cells baseline normal cluster cells}
}
\value{
normalized read counts centered around ploidy 2
}
\description{
Identify ploidy for samples
}
