% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay_object.R
\name{create_assay}
\alias{create_assay}
\title{Create Assay Object}
\usage{
create_assay(assay_name, cell_annotations, feature_annotations)
}
\arguments{
\item{assay_name}{assay name (dna, cnv, protein)}

\item{cell_annotations}{table of data annotating cells, including sample labels}

\item{feature_annotations}{table of data annotating features. i.e. variant names or amplicon name}
}
\value{
Assay object
}
\description{
Create Assay Object
}
\examples{
\dontrun{
genotypes <- extract_genotypes(loom, barcodes, gt.mask=TRUE)
assay <- create_assay(genotypes, "snv", "sample")
}
}
