% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiomics_object.R
\name{create_moo}
\alias{create_moo}
\title{Create Multiomics Object}
\usage{
create_moo(experiment_name, cell_annotations)
}
\arguments{
\item{experiment_name}{Name of Experiment}

\item{cell_annotations}{table of cell annotations. All in this object must match these cell annotations}
}
\value{
Tapestri_Multiomics Object (moo)
}
\description{
Create Multiomics Object
}
