% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay_object.R
\name{subset_assay}
\alias{subset_assay}
\title{Subset assay}
\usage{
subset_assay(assay, keep_cell_ids = TRUE, keep_feature_ids = TRUE)
}
\arguments{
\item{assay}{Assay object to subset}

\item{keep_cell_ids}{vector of cell ids to keep}

\item{keep_feature_ids}{vector of feature ids to keep}
}
\value{
return subsetted assay
}
\description{
Subset assay
}
